[Feature] Log abnormal regex pattern/subject
Summary:
preg_exec can recursively call match() up to a given limit that can be
controlled by the runtime option RuntimeOption::PregRecursionLimit.
The default value is 100000 which is too high and the thread stack will
overflow before the limit is reached. There was a recent incidence of
server crash and we stopped that by setting the limit to 1000. However,
we did not know what regex pattern/subject string combination caused
that crash. Even if not crashing, recursing 1000 times does not seem
to be good for performance. It would be useful to find such abnormal regex
patterns and let application programmers to re-examine a pattern to see if it
can be rewritten. I added some debugging code in the case of two limits
are exceeded: PCRE_EXTRA_MATCH_LIMIT and PCRE_EXTRA_MATCH_LIMIT_RECURSION.
A log message is written to print out the pattern/subject string and some
extra data.

Test Plan:
make fast_tests
build www and set RuntimeOption::PregRecursionLimit low enough to see
such logs are printed.

Reviewed By: mwilliams
Reviewers: qigao, mwilliams
CC: hphp-diffs@lists, ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 258903