[Fix] Make hphpd auto-complete more robust
Summary:
When hphpd is connect to a server, at the server side, an exception can be
thrown while we try to generate the symbols for auto-completion. The exception
is caught and ignored silently. However, we did not finish fetching all the
symbols, and in particular the list of functions isn't setup yet so it is
empty. Therefore the debugger gets an empty list of functions. Auto-complete
failed as the client side did not find any match. I added a try/catch block
to skip the exception (an undeclared constant) and also some logging code
to make the reasoning easier in case similar bug occurs again.

Task ID: # 581390

Blame Rev:

Reviewers:
qigao, mwilliams
CC:
hphp-diffs@lists
Test Plan:
make fast_tests
connect to localhost, and try auto-complete of a function prefix.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 261737