Fix HphpArray copy bug
Summary:
When KindOfIndirect elements are deleted from an HphpArray, the
element entry itself is not removed from the array, instead its data
is decRefed and the data is marked as KindOfUninit.

Later, when this HphpArray is copied, we do not copy the "stale"
KindOfIndirect elements. Instead we create KindOfTombstone elements
in the target array. This had a bug in that the target HphpArray
hash table is copied verbatim from the source, and its hash index
was not being updated to denote that the element is now
KindOfTombstone.

This was causing an assert failure for home.php when we try
to set the array element.

Reviewed By: andrewparoski
Reviewers: kma, je, andrewparoski, oyamauchi
CC: tnovak, ps, mwilliams, andrewparoski, aravind
Revert Plan:
OK

Differential Revision: 267294