Fix bug in hphp_get_iterator
Summary:
hphp_get_iterator() was returning null when its parameter was something
other than an array or an object. This had the unfortunate consequence
of causing a fatal to be thrown if user code tried to call next() on the
value returned by hphp_get_iterator(). Example:

  for ($__foreach__1 = hphp_get_iterator($foo);
       $__foreach__1->valid();
       $__foreach__1->next()) {
    // ...
  }

This is causing problems for yield generators. A normal "foreach" loop
in the body of a yield generator gets transformed into something like the
loop given in the example above, and it will improperly fatal if the value
being iterated over is not an array or object.

This diff fixes the issue and adds a new test to test_code_run.cpp.

Test Plan:
make fast_tests

Reviewed By: je
Reviewers: je, myang, qigao
CC: ps, mwilliams, je, andrewparoski
Differential Revision: 267864