Introduce warnings for defining a constant to be an array or object
Summary:
It's not legal for a constant to have an array or object value in PHP. HPHP was
allowing this to happen, but it was an oversight.

instances of this have cropped up in www, so we can't just kill it off
immediately. This diff adds warnings if you try to do this, but the defines
will
still succeed. Once the instances in www have been cleared out (tasks linked
from #606317), we can make this match Zend behavior; i.e. make the defines
fail.

Test Plan:
make fast_tests, slow_tests

Reviewed By: myang
Reviewers: myang, qigao, andrewparoski, mwilliams
CC: hphp-diffs@lists, ps, mwilliams, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 270159