[Feature] Better stack trace in HPHP
Summary:
HPHP uses a runtime option to control how many frames to include in a PHP stack
trace. When the stack trace contains more frames than the limit, we only
display the top limit frames. Normally the limit should not be reached. When
the limit is reached, often there can be an infinite recursion. I made a
change so that when there are more frames than the limit, we print out
half of the limit from the top of the stack and the other half of limit from
the bottom of the stack.

Test Plan:
make fast_tests

Reviewed By: qigao
Reviewers: qigao
CC: hphp-diffs@lists, ps, mwilliams, qigao, mmachedon, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 271490