[Fix] Prevent use vars in closures from being labelled as un-used, and fix type inference for use vars
Summary:
The VariableTable's definition of a local would consider
use variables in a closure to be a local, causing the dead store
eliminator to wrongfully remove reference assignments in closures
in many cases. This commit fixes that by making a closure use var
not considered as local.

Additionally, type inference for use vars was very sub-optimal,
essentially inferring everything to be a Variant even in simple cases.
This diff fixes this by handling use vars much like parameters are
handled, and using the type information of the use vars in the
surrounding environment to bootstrap the type inference for the closure
itself. The drawback now is that referenced unused use-vars are still declared-
before a benefit of the (sub-optimal) approach was that an un-used use
var never got a declaration set in the symbol table, and thus would not
get emitted. However, I believe that trading better type information for
optimizing for unused referenced use vars is a step forward.

Test Plan:
slow_tests

Reviewed By: mwilliams
Reviewers: mwilliams, myang, qigao, kma
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 271816