[Fix] Fix unitialized class property in generated code
Summary:
When a class is optimized away from the AST, we can miss some processing. In
the new test case, we fail to generator the constructor for the parent class
because it is optimized away from the AST. As a result, we end up with an
uninitialized object property. I moved the relavant processing code to
a place where we do have access to all the user defined classes.

Test Plan:
make fast_tests
make slow_tests
make www and confirm that the generated source code did not change.

Reviewed By: stephentu
Reviewers: stephentu, mwilliams
CC: hphp-diffs@lists, ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 275097