[Fix] $this outside object context should use v_this
Summary:
If $this is used outside object context, and theres a variable table, it gets
added to the table. If the table is writable, the v_this variable can be
modified. So we need to use v_this for non-object context uses of $this in that
case.

<?php
class X {
  function foo($t) {
    $$t = 5;
    var_dump($this);
  }
}
$x = new X;
$x->foo("this");

Test Plan:
fast_tests slow_tests (with new test case).

Reviewed By: myang
Reviewers: stephentu, qigao, myang
Commenters: stephentu
CC: ps, mwilliams, myang, stephentu
Revert Plan:
Tags:

Differential Revision: 275241