[Perf] Implement type assertions and type assertion propagation
Summary:
This diff implements the notion of a type assertion for HPHP's
AST trees. Before the pre-optimize phase of the AliasManager
runs, the AST is walked and type assertions are place in the
AST. Examples of things which trigger type assertions include:

  if (is_array($x)) { /* ARRAY_ASSERT($x) */ }
  if ($x instanceof Foo) { /* OBJECT_ASSERT($x) */ }
  if (is_a($x, Bar)) { /* OBJECT_ASSERT($x) */ }

These type assertions then are used during type-inference
to better propagate type information. For instance, in the
example:

  function f($x) {
    if (is_array($x)) { return $x[0]; }
    return false;
  }

With type assertions we are able to treat the $x in the $x[0]
expression as an array, and we can insert a fast cast for $x.

In addition, this diff modifies the copy-propagator to understand
the notion of type assertions. As a result, we can propagate
type information as such:

  function f($x) {
    if (!is_array($x)) return false;
    // $x now is known to be an array
  }

Test Plan: fast/slow tests, flib tests
Reviewed By: mwilliams
Reviewers: mwilliams
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 276993