[Fix] Using $this in an indirect function call could result in un-compilable code
Summary:
This was an externally reported bug. Using $this in an indirect
function call where we need to stash the function call arguments due to
the possibility of a reference parameter resulted in CPP code which used
v_this to implement $this. This commit fixes this issue by only
outputting v_this if there is an lvalue VariableTable- otherwise, it is
impossible for $this to be modified and we don't need to stash it.

Test Plan: slow_tests
Reviewers: mwilliams, myang, qigao
Commenters: mwilliams
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 277121