[Fix] Do not check/save address of temporary array during serialization
Summary:
When serializing an object, we convert them into an array and call
serialization of the array. The normal serialization rountine for
array would check whether the ArrayData pointer is already serialized
to handle internal reference. The problem is that when serializing an
array converted from an object, the array itself is temporary, and the
ArrayData pointer is likely to be reused later. This could lead to a
mismatch in unserialization.

This diff fixes it by not checking ArrayData pointer if it is temporary
for serializing an object, in which case the ObjectData itself should
take care of the internal reference checking.

Test Plan:
make fast_tests
run the problematic serialization from production and see the object
sequence number mismatch is gone.

Reviewed By: mwilliams
Reviewers: myang, mwilliams
CC: ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 282444
Task ID: 621783