[Cleanup] Try to simplify the handling of symbols during infer types
Summary:
This diff attempts to simplify the handling of symbols for object
properties and constants during type inference. It tries to maintain
several invariants for these kinds of symbols:

  For object properties, the presence of the symbol in the symbol table
  implies the declaration has been set

  For constants, the presence of the symbol in the symbol table implies
  both that the declaration has been set *and* the value has been set

The idea here is that the only thing that should be changing in
type-inference is the value of the types in symbols, not any other
flags, nor should any *new* symbols be created.

This diff is ideally a pre-cursor to a decent implementation of parallel
type-inference.

Finally, this diff fixes a bug where dynamic constants had an actual
type of String, when they should be Variant.

Test Plan: fast_tests, slow_tests, make sure www doesn't change
Reviewed By: mwilliams
Reviewers: mwilliams
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 286862