[Perf/Cleanup] Remove PointerIterator class
Summary:
In SmartAllocatorImpl, the FreeMap class provides an efficient way to find
the next used item. It is prepared by building a dynamic bit map for each and
every item in the allocator. It does so by scanning the free list and for each
item in the free list the corresponding bit is set to 1. The remaining bits
are all zeros representing those used items.
The PointerIterator class gives an abstraction to iterate through each
bit and returns the next item whose bit is not set. However, we can iterate
more directly through the items and move the corresponding bit index in step
without going through the PointerIterator indirection. This is also slightly
faster. So I removed the PointerIterator class, and I renamed prepareIterator
to prepareFreeMap.

Test Plan:
make fast_tests
make slow_tests
make www and do sanity test

Reviewed By: qigao
Reviewers: qigao
CC: hphp-diffs@lists, ps, mwilliams, qigao, myang
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 289196