[Fix] Force a re-run of a user-scope which is currently processing
Summary:
Consider the following case: Scopes A and B are both processing, and
scope A depends on B (say the dependency was added after both A and B
started processing). Now, say scope B finishes with an update that A is
interested in, but scope A is still processing. Since A's current state
could be anywhere in its task, we must assume the worst (that A did not
observe B's update) and re-run scope A. This diff implements this
change.

Test Plan: slow_tests, compile www
Reviewed By: mwilliams
Reviewers: mwilliams, myang, qigao
CC: ps, mwilliams, stephentu
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 293303