Fixed extraordinary DEBUG clowniness
Summary:
We are defining a DEBUG macro in cronoutils.h; this is leaking DEBUG
into the rest of the codebase and causing unrelated ifdef DEBUG gates to
pass. DEBUG is a special compiler flag used for turning on debugging
info! This is so horribly clowny.

In particular, commit 31ff9402 introduced DEBUG gates to mutex.h. When
TAINTED was enabled, this led to the Cronolog Mutex's pthread mutex
being partially overwritten (3 64-bit words from the end; this matches
the size of the four gated values in Mutex (bool, long, int, bool)).

Test Plan:
Ran hphpi server in gdb. DEBUG fields in Mutex are no longer around when
DEBUG is off and the pthread mutex is no longer being clobbered where it
was before.

Reviewed By: mwilliams
Reviewers: mwilliams, myang, stephentu
Commenters: srenfro
CC: ps, mwilliams, srenfro
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 307645