[HPHP Tainting] Copy-on-taint for exposed taint funcs
Summary:
In theory, we ought to maintain copy-on-taint semantics everywhere
since tainting a data object should not necessarily cause all variable
referencing it to acquire the taint. Failing to do so results in funky
behavior like

<?php
$a = 'foo';
$b = $a;
$c = $b;
fb_set_taint($c, TAINT_ALL);
echo fb_get_taint('foo', TAINT_ALL);

returning true. In the normal case, however, we only ever want to create
taints when new strings are created, such was when we taint GET or POST
parameters or pass taint on concatenation. In fact, the only real
exceptions are fb_set_taint() and fb_unset_taint(), which forcibly alter
the taint of a StringData object which could potentially be shared. We
fix this by implementing copy-on-taint in this very narrow case.

Tests that relied on the other semantics have been changed. We also
include some refactoring of test code and a minor fix to taint coverage
in StringData::copy().

Test Plan: TestCodeRun[Eval]::TestTaint
Reviewed By: mwilliams
Reviewers: mwilliams, myang, stephentu
CC: srenfro, amenghra, ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 308351
Task ID: 656564